Ex5/requirements.txt
flask_restfull 
flask_sqlalchemy